// database_repository/db.d.ts
export type Generated<T> = T extends ColumnType<infer S, infer I, infer U>
  ? ColumnType<S, I | undefined, U>
  : ColumnType<T, T | undefined, T>;

export type Json = ColumnType<JsonValue, string, string>;

export type JsonArray = JsonValue[];

export type JsonObject = {
  [x: string]: JsonValue | undefined;
};

export type JsonPrimitive = boolean | number | string | null;

export type JsonValue = JsonArray | JsonObject | JsonPrimitive;

export interface AuthorExternal {
  author_id: number;
  external_id: number;
  id: Generated<number>;
}

export interface AuthorFile {
  author_id: number;
  file_id: number;
  id: Generated<number>;
}

export interface AuthorMembership {
  author_id: number;
  id: Generated<number>;
  is_member_of: number;
}

export interface Authors {
  created_at: Date | null;
  description: string | null;
  email: string | null;
  id: Generated<number>;
  name: string;
  type: Generated<number | null>;
  updated_at: Date | null;
  url: string | null;
  visits: Generated<number>;
}

export interface AuthorSongLyric {
  author_id: number;
  authorship_type: Generated<string>;
  id: Generated<number>;
  song_lyric_id: number;
}

export interface AuthorUser {
  author_id: number;
  created_at: Date | null;
  id: Generated<number>;
  updated_at: Date | null;
  user_id: number;
}

export interface DataRows {
  add: Generated<number>;
  browse: Generated<number>;
  data_type_id: number;
  delete: Generated<number>;
  details: string | null;
  display_name: string;
  edit: Generated<number>;
  field: string;
  id: Generated<number>;
  order: Generated<number>;
  read: Generated<number>;
  required: Generated<number>;
  type: string;
}

export interface DataTypes {
  controller: string | null;
  created_at: Date | null;
  description: string | null;
  details: string | null;
  display_name_plural: string;
  display_name_singular: string;
  generate_permissions: Generated<number>;
  icon: string | null;
  id: Generated<number>;
  model_name: string | null;
  name: string;
  policy_name: string | null;
  server_side: Generated<number>;
  slug: string;
  updated_at: Date | null;
}

export interface ElasticMigrations {
  batch: number;
  migration: string;
}

export interface Externals {
  author_id__obsolete: number | null;
  caption: string | null;
  catalog_number: string | null;
  content_type: Generated<number>;
  copyright: string | null;
  created_at: Date | null;
  editor: string | null;
  has_anonymous_author: Generated<number>;
  id: Generated<number>;
  is_featured: Generated<number>;
  is_uploaded: Generated<number>;
  media_type: string | null;
  published_by: string | null;
  song_lyric_id: number | null;
  updated_at: Date | null;
  url: string | null;
  visits: Generated<string>;
}

export interface FailedJobs {
  connection: string;
  exception: string;
  failed_at: Generated<Date>;
  id: Generated<number>;
  payload: string;
  queue: string;
  uuid: string;
}

export interface Files {
  author_id__obsolete: number | null;
  catalog_number: string | null;
  copyright: string | null;
  created_at: Date | null;
  description: string | null;
  downloads: Generated<number>;
  editor: string | null;
  filename: string;
  has_anonymous_author: Generated<number>;
  id: Generated<number>;
  licence_content: string | null;
  licence_type: number | null;
  name: string | null;
  path: string;
  published_by: string | null;
  song_lyric_id: number | null;
  type: Generated<number>;
  updated_at: Date | null;
}

export interface LilypondPartsSheetMusic {
  created_at: Date | null;
  global_src: string | null;
  id: Generated<number>;
  is_empty: Generated<number>;
  lilypond_parts: Json;
  score_config: Json | null;
  sequence_string: string | null;
  song_lyric_id: number;
  updated_at: Date | null;
}

export interface LiturgicalReferences {
  created_at: Date | null;
  cycle: string | null;
  date: Date;
  id: Generated<number>;
  reading: string | null;
  song_lyric_id: number;
  type: string | null;
  updated_at: Date | null;
}

export interface LiturgicalYearReadings {
  created_at: Date | null;
  date: Date;
  id: Generated<number>;
  reference_1: string | null;
  reference_2: string | null;
  reference_3: string | null;
  reference_type: string;
  references_others: string | null;
  updated_at: Date | null;
  url: string;
}

export interface MenuItems {
  color: string | null;
  created_at: Date | null;
  icon_class: string | null;
  id: Generated<number>;
  menu_id: number | null;
  order: number;
  parameters: string | null;
  parent_id: number | null;
  route: string | null;
  target: Generated<string>;
  title: string;
  updated_at: Date | null;
  url: string;
}

export interface Menus {
  created_at: Date | null;
  id: Generated<number>;
  name: string;
  updated_at: Date | null;
}

export interface Migrations {
  batch: number;
  id: Generated<number>;
  migration: string;
}

export interface ModelHasPermissions {
  model_id: number;
  model_type: string;
  permission_id: number;
}

export interface ModelHasRoles {
  model_id: number;
  model_type: string;
  role_id: number;
}

export interface NewsItems {
  created_at: Date | null;
  expires_at: Date | null;
  fa_icon: string | null;
  id: Generated<number>;
  is_published: Generated<number>;
  link: string;
  link_type: Generated<number>;
  starts_at: Date | null;
  text: string | null;
  updated_at: Date | null;
}

export interface PasswordResets {
  created_at: Date | null;
  email: string;
  token: string;
}

export interface Permissions {
  created_at: Date | null;
  guard_name: string;
  id: Generated<number>;
  name: string;
  updated_at: Date | null;
}

export interface Playlists {
  created_at: Date | null;
  id: Generated<number>;
  is_private: Generated<number>;
  name: string;
  public_user_id: number;
  updated_at: Date | null;
}

export interface PlaylistSongLyric {
  id: Generated<number>;
  order: number;
  playlist_id: number;
  song_lyric_id: number;
}

export interface PublicUsers {
  admin_user_id: number | null;
  created_at: Date | null;
  email: string;
  firebase_id: string;
  id: Generated<number>;
  last_jwt_token: string | null;
  name: string;
  picture_url: string | null;
  updated_at: Date | null;
}

export interface RenderedScores {
  created_at: Date | null;
  external_id: number | null;
  filename: string;
  filetype: string;
  frontend_display_order: number | null;
  id: Generated<number>;
  lilypond_parts_sheet_music_id: number | null;
  render_config: Json | null;
  render_config_hash: string | null;
  secondary_filetypes: Json | null;
  updated_at: Date | null;
}

export interface Revisions {
  created_at: Date | null;
  id: Generated<number>;
  key: string;
  new_value: string | null;
  old_value: string | null;
  revisionable_id: number;
  revisionable_type: string;
  updated_at: Date | null;
  user_id: number | null;
}

export interface RoleHasPermissions {
  permission_id: number;
  role_id: number;
}

export interface Roles {
  created_at: Date | null;
  guard_name: string;
  id: Generated<number>;
  name: string;
  updated_at: Date | null;
}

export interface Settings {
  details: string | null;
  display_name: string;
  group: string | null;
  id: Generated<number>;
  key: string;
  order: Generated<number>;
  type: string;
  value: string | null;
}

export interface SongbookRecords {
  created_at: Date | null;
  id: Generated<number>;
  number: string | null;
  placeholder: string | null;
  song_lyric_id: number | null;
  song_name: string | null;
  songbook_id: number;
  updated_at: Date | null;
}

export interface Songbooks {
  color: string | null;
  color_text: string | null;
  created_at: Date | null;
  id: Generated<number>;
  is_private: Generated<number>;
  name: string;
  shortcut: string | null;
  songbook_img_url: string | null;
  songs_count: number | null;
  updated_at: Date | null;
  updating_at: Date | null;
  updating_user_id: number | null;
}

export interface SongLyricBibleReference {
  book: string;
  end_chapter: number;
  end_verse: number;
  song_lyric_id: number;
  start_chapter: number;
  start_verse: number;
}

export interface SongLyricLilypondSrc {
  id: Generated<number>;
  lilypond_src: string;
  song_lyric_id: number;
}

export interface SongLyricLilypondSvg {
  id: Generated<number>;
  lilypond_svg: string;
  song_lyric_id: number;
}

export interface SongLyricLyrics {
  id: Generated<number>;
  lyrics: string;
  song_lyric_id: number;
}

export interface SongLyrics {
  admin_note: string | null;
  arrangement_of: number | null;
  bible_refs_osis: string | null;
  bible_refs_src: string | null;
  capo: Generated<number>;
  created_at: Date | null;
  deleted_at: Date | null;
  description: string | null;
  has_anonymous_author: Generated<number>;
  has_chords: Generated<number>;
  hymnology: Generated<string>;
  id: Generated<number>;
  is_approved_by_author: Generated<number>;
  is_published: Generated<number>;
  is_sealed: Generated<number>;
  lang: Generated<string>;
  licence_content: string | null;
  licence_type: number | null;
  licence_type_cc: Generated<number>;
  lilypond_key_major: string | null;
  liturgy_approval_status: Generated<number>;
  name: string;
  only_regenschori: Generated<number>;
  revision_n_authors: Generated<number>;
  revision_n_songbook_records: Generated<number>;
  revision_n_tags: Generated<number>;
  secondary_name_1: string | null;
  secondary_name_2: string | null;
  song_id: number | null;
  song_number: number | null;
  type: Generated<number | null>;
  updated_at: Date | null;
  updating_at: Date | null;
  updating_user_id: number | null;
  user_creator_id: number | null;
}

export interface Songs {
  created_at: Date | null;
  deleted_at: Date | null;
  id: Generated<number>;
  name: string;
  updated_at: Date | null;
}

export interface Taggables {
  tag_id: number;
  taggable_id: number;
  taggable_type: string;
}

export interface Tags {
  created_at: Date | null;
  description: string | null;
  hide_in_liturgy: Generated<number>;
  id: Generated<number>;
  lit_day_identificator: string | null;
  name: string;
  order: number | null;
  parent_tag_id: number | null;
  type: Generated<number>;
  updated_at: Date | null;
}

export interface Translations {
  column_name: string;
  created_at: Date | null;
  foreign_key: number;
  id: Generated<number>;
  locale: string;
  table_name: string;
  updated_at: Date | null;
  value: string;
}

export interface Users {
  api_token: string | null;
  avatar: Generated<string | null>;
  created_at: Date | null;
  email: string;
  id: Generated<number>;
  name: string | null;
  password: string | null;
  remember_token: string | null;
  role_id: number | null;
  settings: string | null;
  stats_json: Json | null;
  updated_at: Date | null;
}

export interface VisitAggregates {
  count_after_prune: number;
  count_pruned: Generated<number>;
  count_week: number;
  visitable_id: number;
  visitable_type: string;
}

export interface Visits {
  created_at: Date | null;
  id: Generated<number>;
  is_mobile: number | null;
  public_user_id: number | null;
  source: Generated<number>;
  visit_type: Generated<number>;
  visitable_id: number;
  visitable_type: string;
}

export interface DB {
  author_external: AuthorExternal;
  author_file: AuthorFile;
  author_membership: AuthorMembership;
  author_song_lyric: AuthorSongLyric;
  author_user: AuthorUser;
  authors: Authors;
  data_rows: DataRows;
  data_types: DataTypes;
  elastic_migrations: ElasticMigrations;
  externals: Externals;
  failed_jobs: FailedJobs;
  files: Files;
  lilypond_parts_sheet_music: LilypondPartsSheetMusic;
  liturgical_references: LiturgicalReferences;
  liturgical_year_readings: LiturgicalYearReadings;
  menu_items: MenuItems;
  menus: Menus;
  migrations: Migrations;
  model_has_permissions: ModelHasPermissions;
  model_has_roles: ModelHasRoles;
  news_items: NewsItems;
  password_resets: PasswordResets;
  permissions: Permissions;
  playlist_song_lyric: PlaylistSongLyric;
  playlists: Playlists;
  public_users: PublicUsers;
  rendered_scores: RenderedScores;
  revisions: Revisions;
  role_has_permissions: RoleHasPermissions;
  roles: Roles;
  settings: Settings;
  song_lyric_bible_reference: SongLyricBibleReference;
  song_lyric_lilypond_src: SongLyricLilypondSrc;
  song_lyric_lilypond_svg: SongLyricLilypondSvg;
  song_lyric_lyrics: SongLyricLyrics;
  song_lyrics: SongLyrics;
  songbook_records: SongbookRecords;
  songbooks: Songbooks;
  songs: Songs;
  taggables: Taggables;
  tags: Tags;
  translations: Translations;
  users: Users;
  visit_aggregates: VisitAggregates;
  visits: Visits;
}